Use u64 for R2 range requests
Closes #560

+++ worker-sys/src/types/r2/range.rs
@@ -3,7 +3,7 @@ use wasm_bindgen::prelude::*;
 #[wasm_bindgen]
 #[derive(Debug, Clone)]
 pub struct R2Range {
-    pub offset: Option<u32>,
-    pub length: Option<u32>,
-    pub suffix: Option<u32>,
+    pub offset: Option<f64>,
+    pub length: Option<f64>,
+    pub suffix: Option<f64>,
 }
